moved login auth back into routes template
moved the login auth processing for @login_manager back to routes.py, resolved issue reporting 'TypeError: 'Response' object is not iterable'
